+++ test/transform/resource/after-delombok/LoggerXSlf4j.java
@@ -1,0 +1,7 @@
+class LoggerXSlf4j {
+	private static final org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXSlf4j.class);
+}
+
+class LoggerXSlf4jWithImport {
+	private static final org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXSlf4jWithImport.class);
+}