+++ test/transform/resource/after-ecj/LoggerXSlf4j.java
@@ -1,0 +1,17 @@
+import lombok.extern.slf4j.XSlf4j;
+@lombok.extern.slf4j.XSlf4j class LoggerXSlf4j {
+  private static final org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXSlf4j.class);
+  <clinit>() {
+  }
+  LoggerXSlf4j() {
+    super();
+  }
+}
+@XSlf4j class LoggerXSlf4jWithImport {
+  private static final org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXSlf4jWithImport.class);
+  <clinit>() {
+  }
+  LoggerXSlf4jWithImport() {
+    super();
+  }
+}